LayoutTests/fast/dom/MutationObserver/parser-mutations-expected.txt

PASS mutations.length is 6
PASS mutations[1].type is "childList"
PASS mutations[1].target.tagName is "BODY"
PASS mutations[1].addedNodes.length is 1
PASS mutations[1].addedNodes[0].tagName is "P"
Mutation records should be delivered for all parser mutations after the above script.
